requirements.txt
openai
python-dotenv
pandas
numpy
psycopg2
pymupdf
fastapi
uvicorn
pypdf